// packages/ui/src/components/ui/form_disabled/form.types.ts
export interface FormLabelProps {
  for?: string
  required?: boolean
  optional?: boolean
  validationState?: ValidationState
  tooltip?: JSX.Element
  icon?: JSX.Element
  class?: string
  children: JSX.Element
}